Remove suppression_mappings.txt (#2229)
* Fix all suppressed -Wunused-variable

* Fix all suppressed -Wimplicit-int-conversion

* Fix all suppressed -Wunused-const-variable

* Fix all suppressed -Wsign-compare

* Fix all suppressed -Wshorten-64-to-32

* Fix all suppressed -Wreturn-type

* Suppress -Wimplicit-int-float-conversion inside rocksdb/db.h inline

A better solution will be to upstream a fix but in the meantime, this
will be good enough.

Since this is the last suppressed warning, the commit also deletes
suppression_mappings.txt and CMake glue for it.

* Remove annoying stdout output from asm-to-cpp.fif

* Remove stale files from null consensus

+++ validator/full-node-fast-sync-overlays.cpp
@@ -514,7 +514,7 @@ void FullNodeFastSyncOverlays::update_overlays(
   for (const auto &desc : state->get_shards()) {
     ShardIdFull shard = desc->shard();
     td::uint32 monitor_min_split = state->monitor_min_split_depth(shard.workchain);
-    if (shard.pfx_len() > monitor_min_split) {
+    if (shard.pfx_len() > static_cast<int>(monitor_min_split)) {
       shard = shard_prefix(shard, monitor_min_split);
     }
     all_shards.insert(shard);
